+++ packages/slate-linkify/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # [0.2.0] - 2020-01-12
 
-- refactored to fit slate@0.50.0 and above
+- [BREAKING CHANGES]: refactored to fit slate@0.50.0 and above
 
 # [0.1.2] - 2019-03-22